requirements.txt
backtrader==1.9.76.123
Flask==1.1.2
matplotlib~=3.2.1
nsepy==0.8
numpy~=1.18.4
pandas~=1.0.3
pyfolio==0.9.2
python-crontab==2.5.1
python-dateutil==2.8.1
scipy==1.4.1
seaborn~=0.11.0
statsmodels==0.11.1
trendet==0.7
trendln==0.1.10

mplfinance~=0.12.7a7
